更新

+++ goods/src/main/java/cn/edu/xmu/goods/controller/BrandController.java
@@ -17,6 +17,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 import springfox.documentation.annotations.ApiIgnore;
 
 import javax.servlet.http.HttpServletResponse;
@@ -35,14 +36,19 @@ public class BrandController {
     private HttpServletResponse httpServletResponse;
 
     @DeleteMapping("/shops/{shopId}/brands/{id}")
-    public Object deleteBrand(@PathVariable("id") Long id){
-        ReturnObject returnObject = brandService.deleteBrand(id);
-        return Common.decorateReturnObject(returnObject);
+    public Object deleteBrand(@PathVariable("shopId") Long shopId, @PathVariable("id") Long id,
+                              @Depart @ApiIgnore @RequestParam(required = false) Long departId){
+        departId= Long.valueOf(0);//平台
+        if(shopId.equals(departId)) {
+            ReturnObject returnObject = brandService.deleteBrand(id);
+            return Common.decorateReturnObject(returnObject);
+        }
+        else return Common.getNullRetObj(new ReturnObject<>(ResponseCode.RESOURCE_ID_OUTSCOPE,String.format("无权限操作：" + shopId)), httpServletResponse);
     }
 
 
     /**
-     * 修改角色信息
+     * 修改品牌信息
      *
      * @author 24320182203281 王纬策
      * @param id 角色id
@@ -52,7 +58,7 @@ public class BrandController {
      * createdBy 王纬策 2020/11/04 13:57
      * modifiedBy 王纬策 2020/11/7 19:20
      */
-    @ApiOperation(value = "修改角色信息", produces = "application/json")
+    @ApiOperation(value = "修改品牌信息", produces = "application/json")
     @ApiImplicitParams({
             @ApiImplicitParam(paramType = "header", dataType = "String", name = "authorization", value = "Token", required = true),
             @ApiImplicitParam(paramType = "body", dataType = "RoleVo", name = "vo", value = "可修改的用户信息", required = true)
@@ -65,17 +71,17 @@ public class BrandController {
         if (null != returnObject) {
             return returnObject;
         }
-        DepartId= Long.valueOf(0);
+        DepartId= Long.valueOf(0);//平台
         if(shopId.equals(DepartId)){
             Brand brand = vo.createBrand();
             brand.setId(id);
-
+            brand.setGmtModified(LocalDateTime.now());
             ReturnObject<Object> retObject = brandService.setBrand(id,vo);
             httpServletResponse.setStatus(HttpStatus.OK.value());
             return Common.decorateReturnObject(retObject);
         }
         else{
-            return Common.getNullRetObj(new ReturnObject<>(ResponseCode.FIELD_NOTVALID, String.format("店铺id不匹配：" + shopId)), httpServletResponse);
+            return Common.getNullRetObj(new ReturnObject<>(ResponseCode.RESOURCE_ID_OUTSCOPE,String.format("无权限操作：" + shopId)), httpServletResponse);
         }
 
 
@@ -91,8 +97,8 @@ public class BrandController {
      * createdBy 徐悦 2020/11/20 23：39
      */
     @ApiOperation(value = "新增品牌", produces = "application/json")
-    @PostMapping("/shops/{id}/brands")
-    public Object insertBrand(@Validated @RequestBody BrandVo vo, BindingResult bindingResult,
+    @PostMapping("/shops/{shopId}/brands")
+    public Object insertBrand(@PathVariable("shopId") Long shopId,@Validated @RequestBody BrandVo vo, BindingResult bindingResult,
                              @LoginUser @ApiIgnore @RequestParam(required = false) Long userId,
                              @Depart @ApiIgnore @RequestParam(required = false) Long departId) {
         //校验前端数据
@@ -100,10 +106,13 @@ public class BrandController {
         if (null != returnObject) {
             return returnObject;
         }
-
-        ReturnObject<VoObject> retObject = brandService.addBrand(vo);
-        httpServletResponse.setStatus(HttpStatus.OK.value());
-        return Common.decorateReturnObject(retObject);
+        departId= Long.valueOf(0);//平台
+        if(shopId.equals(departId)) {
+            ReturnObject<VoObject> retObject = brandService.addBrand(vo);
+            httpServletResponse.setStatus(HttpStatus.OK.value());
+            return Common.decorateReturnObject(retObject);
+        }
+        else return Common.getNullRetObj(new ReturnObject<>(ResponseCode.RESOURCE_ID_OUTSCOPE,String.format("无权限操作：" + shopId)), httpServletResponse);
     }
 
 
@@ -116,4 +125,23 @@ public class BrandController {
         ReturnObject<PageInfo<VoObject>> returnObject =  brandService.selectAllBrands(page, pageSize);
         return Common.getPageRetObject(returnObject);
     }
+    @ApiOperation(value = "新增品牌图片", produces = "application/json")
+    @PostMapping("/shops/{shopId}/brands/{id}/uploadImg")
+    public Object uploadimage(@PathVariable("shopId") Long shopId, @PathVariable("id") Long id, @RequestParam("img") MultipartFile multipartFile, @Validated @RequestBody BindingResult bindingResult,
+                              @LoginUser @ApiIgnore @RequestParam(required = false) Long userId,
+                              @Depart @ApiIgnore @RequestParam(required = false) Long departId) {
+        //校验前端数据
+        Object returnObject = Common.processFieldErrors(bindingResult, httpServletResponse);
+        if (null != returnObject) {
+            return returnObject;
+        }
+        departId= Long.valueOf(0);//平台
+        if(shopId.equals(departId)) {
+            ReturnObject<Object> retObject = brandService.uploadimage(id,multipartFile);
+            httpServletResponse.setStatus(HttpStatus.OK.value());
+            return Common.decorateReturnObject(retObject);
+        }
+        else return Common.getNullRetObj(new ReturnObject<>(ResponseCode.RESOURCE_ID_OUTSCOPE,String.format("无权限操作：" + shopId)), httpServletResponse);
+    }
+
 }
